pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "debriefeur"
version = "0.1.0"
description = "Debriefeur — extract your expertise into AI agent configurations"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "AIgenteur" }]
license = { text = "MIT" }
dependencies = [
    # LLM — provider-agnostic
    "litellm>=1.55.0,<2",

    # CLI interface
    "rich>=14.0.0,<15",
    "prompt_toolkit>=3.0.50,<4",

    # Data validation
    "pydantic>=2.10.0,<3",

    # HTTP client
    "httpx>=0.28.0,<1",

    # Environment
    "python-dotenv>=1.0.1,<2",

    # Config output
    "pyyaml>=6.0.2,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.24.0,<1",
    "ruff>=0.8.0,<1",
]

[project.scripts]
debriefeur = "debriefeur.cli:main"

[tool.setuptools.packages.find]
include = ["debriefeur*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
